rename: map.ts 파일이름 타입이름과 동일하도록 변경
